+++ src/geefetch/data/get.py
@@ -10,7 +10,7 @@ from rasterio.crs import CRS
 from retry import retry
 from rich.progress import Progress
 
-from ..utils.enums import CompositeMethod, DType, Format, P2Orbit, S1Orbit
+from ..utils.enums import CompositeMethod, DType, Format, P2Orbit, ResamplingMethod, S1Orbit
 from ..utils.progress import default_bar
 from ..utils.rasterio import create_vrt
 from .downloadables import DownloadableABC
@@ -567,6 +567,7 @@ def download_s1(
     dtype: DType = DType.Float32,
     filter_polygon: shapely.Geometry | None = None,
     orbit: S1Orbit = S1Orbit.ASCENDING,
+    resampling: ResamplingMethod = ResamplingMethod.BILINEAR,
 ) -> None:
     """Download Sentinel-1 images. Images are written in several .tif chips
     to `data_dir`. Additionally, a file `s1.vrt` is written to combine all the chips.
@@ -603,6 +604,10 @@ def download_s1(
         More fine-grained AOI than `bbox`. Defaults to None.
     orbit : S1Orbit
         The orbit used to filter Sentinel-1 images. Defaults to S1Orbit.ASCENDING.
+    resampling : ResamplingMethod
+        The resampling method to use when reprojecting images.
+        Can be BILINEAR, BICUBIC or NEAREST.
+        Defaults to ResamplingMethod.BILINEAR.
     """
     download_func = (
         download_time_series if (composite_method == CompositeMethod.TIMESERIES) else download
@@ -638,6 +643,8 @@ def download_s1(
             "dtype": dtype,
             "orbit": orbit,
             "selected_bands": selected_bands,
+            "resampling": resampling,
+            "resolution": resolution,
         },
         satellite_download_kwargs={"dtype": dtype.to_str()},
     )
@@ -658,6 +665,7 @@ def download_s2(
     filter_polygon: shapely.Geometry | None = None,
     cloudless_portion: int = 60,
     cloud_prb_thresh: int = 40,
+    resampling: ResamplingMethod = ResamplingMethod.BILINEAR,
 ) -> None:
     """Download Sentinel-2 images. Images are written in several .tif chips
     to `data_dir`. Additionally, a file `s2.vrt` is written to combine all the chips.
@@ -697,6 +705,10 @@ def download_s2(
         See :meth:`geefetch.data.s2.get`. Defaults to 60.
     cloud_prb_thresh : int
         Cloud probability threshold. See :meth:`geefetch.data.s2.get`. Defaults to 40.
+    resampling : ResamplingMethod
+        The resampling method to use when reprojecting images.
+        Can be BILINEAR, BICUBIC or NEAREST.
+        Defaults to ResamplingMethod.BILINEAR.
     """
     download_func = (
         download_time_series if composite_method == CompositeMethod.TIMESERIES else download
@@ -720,6 +732,8 @@ def download_s2(
             "cloudless_portion": cloudless_portion,
             "cloud_prb_thresh": cloud_prb_thresh,
             "dtype": dtype,
+            "resampling": resampling,
+            "resolution": resolution,
         },
         satellite_download_kwargs={"dtype": dtype.to_str()},
     )
@@ -738,6 +752,7 @@ def download_dynworld(
     composite_method: CompositeMethod = CompositeMethod.MEDIAN,
     dtype: DType = DType.Float32,
     filter_polygon: shapely.Geometry | None = None,
+    resampling: ResamplingMethod = ResamplingMethod.BILINEAR,
 ) -> None:
     """Download Dynamic World images. Images are written in several .tif chips
     to `data_dir`. Additionnally a file `dynworld.vrt` is written to combine all the chips.
@@ -772,6 +787,10 @@ def download_dynworld(
         The data type of the downloaded images. Defaults to DType.Float32.
     filter_polygon : shapely.Geometry | None
         More fine-grained AOI than `bbox`. Defaults to None.
+    resampling : ResamplingMethod
+        The resampling method to use when reprojecting images.
+        Can be BILINEAR, BICUBIC or NEAREST.
+        Defaults to ResamplingMethod.BILINEAR.
     """
     download_func = (
         download_time_series if composite_method == CompositeMethod.TIMESERIES else download
@@ -793,6 +812,8 @@ def download_dynworld(
         satellite_get_kwargs={
             "composite_method": composite_method,
             "dtype": dtype,
+            "resampling": resampling,
+            "resolution": resolution,
         },
         satellite_download_kwargs={"dtype": dtype.to_str()},
     )
@@ -811,6 +832,7 @@ def download_landsat8(
     composite_method: CompositeMethod = CompositeMethod.MEDIAN,
     dtype: DType = DType.Float32,
     filter_polygon: shapely.Geometry | None = None,
+    resampling: ResamplingMethod = ResamplingMethod.BILINEAR,
 ) -> None:
     """Download Landsat 8 images. Images are written in several .tif chips
     to `data_dir`. Additionally, a file `landsat8.vrt` is written to combine all the chips.
@@ -845,6 +867,10 @@ def download_landsat8(
         The data type of the downloaded images. Defaults to DType.Float32.
     filter_polygon : shapely.Geometry | None
         More fine-grained AOI than `bbox`. Defaults to None.
+    resampling : ResamplingMethod
+        The resampling method to use when reprojecting images.
+        Can be BILINEAR, BICUBIC or NEAREST.
+        Defaults to ResamplingMethod.BILINEAR.
     """
     download_func = (
         download_time_series if composite_method == CompositeMethod.TIMESERIES else download
@@ -864,6 +890,8 @@ def download_landsat8(
         satellite_get_kwargs={
             "composite_method": composite_method,
             "dtype": dtype,
+            "resampling": resampling,
+            "resolution": resolution,
         },
         satellite_download_kwargs={"dtype": dtype.to_str()},
     )
@@ -883,6 +911,8 @@ def download_palsar2(
     dtype: DType = DType.Float32,
     filter_polygon: shapely.Geometry | None = None,
     orbit: P2Orbit = P2Orbit.DESCENDING,
+    resampling: ResamplingMethod = ResamplingMethod.BILINEAR,
+    refined_lee: bool = True,
 ) -> None:
     """Download Palsar 2 images. Images are written in several .tif chips
     to `data_dir`. Additionally, a file `palsar2.vrt` is written to combine all the chips.
@@ -919,7 +949,13 @@ def download_palsar2(
         More fine-grained AOI than `bbox`. Defaults to None.
     orbit : P2Orbit
         The orbit used to filter Palsar-2 images. Defaults to P2Orbit.ASCENDING.
-
+    resampling : ResamplingMethod
+        The resampling method to use when reprojecting images.
+        Can be BILINEAR, BICUBIC or NEAREST.
+        Defaults to ResamplingMethod.BILINEAR.
+    refined_lee : bool
+        Whether to apply the Refined Lee filter to reduce speckle noise.
+        Defaults to True.
     """
     download_func = (
         download_time_series if composite_method == CompositeMethod.TIMESERIES else download
@@ -940,6 +976,9 @@ def download_palsar2(
             "composite_method": composite_method,
             "dtype": dtype,
             "orbit": orbit,
+            "resampling": resampling,
+            "resolution": resolution,
+            "refined_lee": refined_lee,
         },
         satellite_download_kwargs={"dtype": dtype.to_str()},
     )
@@ -956,6 +995,7 @@ def download_nasadem(
     composite_method: CompositeMethod = CompositeMethod.MEDIAN,
     dtype: DType = DType.Float32,
     filter_polygon: shapely.Polygon | None = None,
+    resampling: ResamplingMethod = ResamplingMethod.BILINEAR,
 ) -> None:
     """Download NASADEM images. Images are written in several .tif chips
     to `data_dir`. Additionally, a file `nasadem.vrt` is written to combine all the chips.
@@ -986,6 +1026,10 @@ def download_nasadem(
         The data type of the downloaded images. Defaults to DType.Float32.
     filter_polygon : shapely.Polygon | None
         More fine-grained AOI than `bbox`. Defaults to None.
+    resampling : ResamplingMethod
+        The resampling method to use when reprojecting images.
+        Can be BILINEAR, BICUBIC or NEAREST.
+        Defaults to ResamplingMethod.BILINEAR.
     """
     if composite_method == CompositeMethod.TIMESERIES:
         raise ValueError("Time series is not relevant for DEM.")
@@ -1006,6 +1050,7 @@ def download_nasadem(
         satellite_get_kwargs={
             "composite_method": composite_method,
             "dtype": dtype,
+            "resampling": resampling,
         },
         satellite_download_kwargs={"dtype": dtype.to_str()},
     )
@@ -1025,6 +1070,7 @@ def download_custom(
     composite_method: CompositeMethod = CompositeMethod.MEDIAN,
     dtype: DType = DType.Float32,
     filter_polygon: shapely.Polygon | None = None,
+    resampling: ResamplingMethod = ResamplingMethod.BILINEAR,
 ) -> None:
     """Download images from a custom data source. Images are written in several .tif chips
     to `data_dir`. Additionally, a file `nasadem.vrt` is written to combine all the chips.
@@ -1060,6 +1106,10 @@ def download_custom(
         The data type of the downloaded images. Defaults to DType.Float32.
     filter_polygon : shapely.Polygon | None
         More fine-grained AOI than `bbox`. Defaults to None.
+    resampling : ResamplingMethod
+        The resampling method to use when reprojecting images.
+        Can be BILINEAR, BICUBIC or NEAREST.
+        Defaults to ResamplingMethod.BILINEAR.
     """
     if composite_method == CompositeMethod.TIMESERIES:
         raise ValueError("Time series is not relevant for Custom Satellites.")
@@ -1080,6 +1130,8 @@ def download_custom(
         satellite_get_kwargs={
             "composite_method": composite_method,
             "dtype": dtype,
+            "resampling": resampling,
+            "resolution": resolution,
         },
         satellite_download_kwargs={"dtype": dtype.to_str()},
     )