+++ src/geefetch/data/satellites/dem.py
@@ -5,8 +5,7 @@ from ee.image import Image
 from ee.terrain import Terrain
 from geobbox import GeoBoundingBox
 
-from ...utils.enums import CompositeMethod, DType
-from ...utils.rasterio import WGS84
+from ...utils.enums import CompositeMethod, DType, ResamplingMethod
 from ..downloadables import DownloadableGeedimImage, DownloadableGeedimImageCollection
 from ..downloadables.geedim import PatchedBaseImage
 from .abc import SatelliteABC
@@ -73,6 +72,8 @@ class NASADEM(SatelliteABC):
         end_date: str | None = None,
         composite_method: CompositeMethod = CompositeMethod.MEAN,
         dtype: DType = DType.Float32,
+        resampling: ResamplingMethod = ResamplingMethod.BILINEAR,
+        resolution: float = 30,
         **kwargs: Any,
     ) -> DownloadableGeedimImage:
         """Get a downloadable NASADEM composite image.
@@ -89,6 +90,10 @@ class NASADEM(SatelliteABC):
             (Unused) NASADEM is a single static dataset.
         dtype : DType
             The data type for the image.
+        resampling : ResamplingMethod
+            The resampling method to use when processing the image.
+        resolution : float
+            The resolution for the image.
         **kwargs : Any
             Accepted but ignored additional arguments.
 
@@ -100,9 +105,11 @@ class NASADEM(SatelliteABC):
         for key in kwargs:
             log.warning(f"Argument {key} is ignored.")
 
-        bounds = aoi.transform(WGS84).to_ee_geometry()
-        dem_im = self.get_im().clip(bounds)
-        dem_im = self.convert_image(dem_im, dtype)
+        # resample
+        dem_im = self.get_im()
+        dem_im = self.resample_reproject_clip(dem_im, aoi, resampling, resolution)
+        # apply dtype
+        dem_im = self.convert_dtype(dem_im, dtype)
         return DownloadableGeedimImage(PatchedBaseImage(dem_im))
 
     @property