+++ src/geefetch/data/satellites/palsar2.py
@@ -1,13 +1,14 @@
 import logging
 from typing import Any
 
+import ee
 from ee.filter import Filter
 from ee.image import Image
 from ee.imagecollection import ImageCollection
 from geobbox import GeoBoundingBox
 from shapely import Polygon
 
-from ...utils.enums import CompositeMethod, DType, P2Orbit
+from ...utils.enums import CompositeMethod, DType, P2Orbit, ResamplingMethod
 from ...utils.rasterio import WGS84
 from ..downloadables import DownloadableGeedimImage, DownloadableGeedimImageCollection
 from ..downloadables.geedim import PatchedBaseImage
@@ -40,7 +41,7 @@ class Palsar2(SatelliteABC):
 
     @property
     def pixel_range(self):
-        return 0, 8000
+        return -30, 0  # log10 scale
 
     @property
     def resolution(self):
@@ -91,7 +92,10 @@ class Palsar2(SatelliteABC):
         start_date: str | None = None,
         end_date: str | None = None,
         dtype: DType = DType.Float32,
+        resampling: ResamplingMethod = ResamplingMethod.BILINEAR,
         orbit: P2Orbit = P2Orbit.DESCENDING,
+        resolution: float = 25,
+        refined_lee: bool = True,
         **kwargs: Any,
     ) -> DownloadableGeedimImageCollection:
         """Get a downloadable time series of Palsar-2 images.
@@ -106,8 +110,14 @@ class Palsar2(SatelliteABC):
             End date in "YYYY-MM-DD" format.
         dtype : DType
             The data type for the image
+        resampling : ResamplingMethod
+            The resampling method to use when processing the image.
         orbit : P2Orbit
             The orbit used to filter the collection before mosaicking.
+        resolution: float
+            The resolution for the image.
+        refined_lee : bool
+            Whether to apply the Refined Lee filter to reduce speckle noise.
         **kwargs : Any
             Accepted but ignored additional arguments.
 
@@ -118,20 +128,21 @@ class Palsar2(SatelliteABC):
         """
         p2_col = self.get_col(aoi, start_date, end_date, orbit)
 
-        images = {}
+        # get the info of the collection
         info = p2_col.getInfo()
         n_images = len(info["features"])  # type: ignore[index]
         if n_images == 0:
             log.error(f"Found 0 Palsar-2 image." f"Check region {aoi.transform(WGS84)}.")
             raise RuntimeError("Collection of 0 Palsar-2 image.")
+        images = {}
         for feature in info["features"]:  # type: ignore[index]
             id_ = feature["id"]
             footprint = PatchedBaseImage.from_id(id_).footprint
             assert footprint is not None
-            if Polygon(footprint["coordinates"][0]).intersects(aoi.to_shapely_polygon()):
-                # aoi intersects im
+            if Polygon(footprint["coordinates"][0]).contains(aoi.to_shapely_polygon()):
                 im = Image(id_)
-                im = self.convert_image(im, dtype)
+                im = self.before_composite(im, resampling, aoi, resolution, refined_lee)
+                im = self.after_composite(im, dtype)
                 images[id_.removeprefix("JAXA/ALOS/PALSAR-2/Level2_2/ScanSAR/")] = PatchedBaseImage(
                     im
                 )
@@ -144,7 +155,10 @@ class Palsar2(SatelliteABC):
         end_date: str | None = None,
         composite_method: CompositeMethod = CompositeMethod.MEAN,
         dtype: DType = DType.Float32,
+        resampling: ResamplingMethod = ResamplingMethod.BILINEAR,
         orbit: P2Orbit = P2Orbit.DESCENDING,
+        resolution: float = 25,
+        refined_lee: bool = True,
         **kwargs: Any,
     ) -> DownloadableGeedimImage:
         """Get a downloadable mosaic of Palsar-2 images.
@@ -161,8 +175,14 @@ class Palsar2(SatelliteABC):
             The method use to do mosaicking.
         dtype : DType
             The data type for the image
+        resampling : ResamplingMethod
+            The resampling method to use when processing the image.
         orbit : P2Orbit
             The orbit used to filter the collection before mosaicking
+        resolution: float
+            The resolution for the image.
+        refined_lee : bool
+            Whether to apply the Refined Lee filter to reduce speckle noise.
         **kwargs : Any
             Accepted but ignored additional arguments.
 
@@ -174,9 +194,7 @@ class Palsar2(SatelliteABC):
         for key in kwargs:
             log.warning(f"Argument {key} is ignored.")
 
-        bounds = aoi.transform(WGS84).to_ee_geometry()
         p2_col = self.get_col(aoi, start_date, end_date, orbit)
-
         info = p2_col.getInfo()
         n_images = len(info["features"])  # type: ignore
         if n_images > 500:
@@ -187,10 +205,13 @@ class Palsar2(SatelliteABC):
         if n_images == 0:
             log.error(f"Found 0 Palsar-2 image." f"Check region {aoi.transform(WGS84)}.")
             raise RuntimeError("Collection of 0 Palsar-2 image.")
-
         log.debug(f"Palsar-2 mosaicking with {n_images} images.")
+        p2_col = p2_col.map(
+            lambda img: self.before_composite(img, resampling, aoi, resolution, refined_lee)
+        )
+        bounds = aoi.transform(WGS84).to_ee_geometry()
         p2_im = composite_method.transform(p2_col).clip(bounds)
-        p2_im = self.convert_image(p2_im, dtype)
+        p2_im = self.after_composite(p2_im, dtype)
         p2_im = PatchedBaseImage(p2_im)
         return DownloadableGeedimImage(p2_im)
 
@@ -201,3 +222,53 @@ class Palsar2(SatelliteABC):
     @property
     def full_name(self) -> str:
         return "Palsar-2"
+
+    def before_composite(
+        self,
+        im: Image,
+        resampling: ResamplingMethod,
+        aoi: GeoBoundingBox,
+        scale: float,
+        refined_lee: bool = False,
+    ) -> Image:
+        # Convert from DN to power
+        im = im.pow(2)
+        # Optionally apply a refined lee filter
+        if refined_lee:
+            im = _refined_lee(im)
+        # Apply resampling if specified
+        im = self.resample_reproject_clip(im, aoi, resampling, scale)
+        return im
+
+    def after_composite(self, im: Image, dtype: DType, log_scale: bool = True) -> Image:
+        # Convert from power to gamma0:
+        if log_scale:
+            im = im.log10().multiply(10).subtract(83)
+        # Apply pixel range and dtype
+        im = self.convert_dtype(im, dtype)
+        return im
+
+
+def _refined_lee(image: Image) -> Image:
+    """
+    Apply the Refined Lee filter to reduce speckle noise.
+    Parameters
+    ----------
+    image : Image
+        The input image to be filtered.
+    Returns
+    -------
+    Image
+        The image with the Refined Lee filter applied.
+    """
+
+    def apply_filter(band_name: str) -> Image:
+        band = image.select(band_name)
+        mean = band.reduceNeighborhood(ee.Reducer.mean(), ee.Kernel.square(3))
+        variance = band.reduceNeighborhood(ee.Reducer.variance(), ee.Kernel.square(3))
+        weight = variance.divide(variance.add(mean.pow(2)))
+        return mean.add(weight.multiply(band.subtract(mean))).rename(band_name)  # type: ignore[no-any-return]
+
+    filtered_hh = apply_filter("HH")
+    filtered_hv = apply_filter("HV")
+    return image.addBands([filtered_hh, filtered_hv], overwrite=True)