+++ src/geefetch/data/satellites/s1.py
@@ -1,13 +1,14 @@
 import logging
 from typing import Any
 
+import ee
 from ee.filter import Filter
 from ee.image import Image
 from ee.imagecollection import ImageCollection
 from geobbox import GeoBoundingBox
 from shapely import Polygon
 
-from ...utils.enums import CompositeMethod, DType, S1Orbit
+from ...utils.enums import CompositeMethod, DType, ResamplingMethod, S1Orbit
 from ...utils.rasterio import WGS84
 from ..downloadables import DownloadableGeedimImage, DownloadableGeedimImageCollection
 from ..downloadables.geedim import PatchedBaseImage
@@ -120,6 +121,8 @@ class S1(SatelliteABC):
         dtype: DType = DType.Float32,
         orbit: S1Orbit = S1Orbit.ASCENDING,
         selected_bands: list[str] | None = None,
+        resampling: ResamplingMethod = ResamplingMethod.BILINEAR,
+        resolution: float = 10,
         **kwargs: Any,
     ) -> DownloadableGeedimImageCollection:
         """Get a downloadable time series of Sentinel-1 images.
@@ -138,6 +141,10 @@ class S1(SatelliteABC):
             The orbit used to filter the collection before mosaicking
         selected_bands : list[str] | None
             The bands to be downloaded.
+        resampling : ResamplingMethod
+            The resampling method to use when compositing images.
+        resolution: float
+            The resolution for the image.
         **kwargs : Any
             Accepted but ignored additional arguments.
 
@@ -163,7 +170,10 @@ class S1(SatelliteABC):
             if Polygon(footprint["coordinates"][0]).intersects(aoi.to_shapely_polygon()):
                 # aoi intersects im
                 im = Image(id_)
-                im = self.convert_image(im, dtype)
+                # convert to power and resample
+                im = self.before_composite(im, resampling, aoi, resolution)
+                # Apply pixel range and dtype
+                im = self.after_composite(im, dtype)
                 images[id_.removeprefix("COPERNICUS/S1_GRD/")] = PatchedBaseImage(im)
         return DownloadableGeedimImageCollection(images)
 
@@ -176,6 +186,8 @@ class S1(SatelliteABC):
         dtype: DType = DType.Float32,
         orbit: S1Orbit = S1Orbit.ASCENDING,
         selected_bands: list[str] | None = None,
+        resampling: ResamplingMethod = ResamplingMethod.BILINEAR,
+        resolution: float = 10,
         **kwargs: Any,
     ) -> DownloadableGeedimImage:
         """Get a downloadable mosaic of Sentinel-1 images.
@@ -196,6 +208,10 @@ class S1(SatelliteABC):
             The orbit used to filter the collection before mosaicking
         selected_bands : list[str] | None
             The bands to be downloaded.
+        resampling : ResamplingMethod
+            The resampling method to use when compositing images.
+        resolution: float
+            The resolution for the image.
         **kwargs : Any
             Accepted but ignored additional arguments.
 
@@ -222,9 +238,14 @@ class S1(SatelliteABC):
                 raise RuntimeError("Collection of 0 Sentinel-1 image.")
 
             log.debug(f"Sentinel-1 mosaicking with {n_images} images.")
+
+            # Process all images in the collection
+            s1_col = s1_col.map(lambda im: self.before_composite(im, resampling, aoi, resolution))
+            # Create composite
             bounds = aoi.transform(WGS84).to_ee_geometry()
             s1_im = composite_method.transform(s1_col).clip(bounds)
-            s1_im = self.convert_image(s1_im, dtype)
+            # Process composite
+            s1_im = self.after_composite(s1_im, dtype)
             return s1_im
 
         match orbit:
@@ -241,6 +262,26 @@ class S1(SatelliteABC):
 
         s1_im = PatchedBaseImage(s1_im)
         return DownloadableGeedimImage(s1_im)
+
+    def before_composite(
+        self,
+        im: Image,
+        resampling: ResamplingMethod,
+        aoi: GeoBoundingBox,
+        scale: float,
+    ) -> Image:
+        # Convert from db to power: 10^(im/10)
+        im = ee.Image(10).pow(im.divide(10))
+        # Apply resampling if specified
+        im = self.resample_reproject_clip(im, aoi, resampling, scale)
+        return im
+
+    def after_composite(self, im: Image, dtype: DType) -> Image:
+        # Convert from power to db
+        im = im.log10().multiply(10)
+        # Apply pixel range and dtype
+        im = self.convert_dtype(im, dtype)
+        return im
 
     @property
     def name(self) -> str: